moved utils.py to exceptions.py
